Removed code duplication between server previewGeneration.js and client EmbedFrame.jsx. #2805
